Preserve mirror schema compatibility
Use i64 for signed delays without narrowing the old u32 range. Accept legacy duration_average input while serializing the canonical duration_avg field, with regression coverage for both forms and the preserved delay range.

Amp-Thread-ID: https://ampcode.com/threads/T-01a0d2f8-70bd-7429-a7e2-a6e8a9c0511e

+++ crates/arch-reflector/src/main.rs
@@ -595,7 +595,7 @@ fn filter_status(filters: &Filters, status: &mut Status) {
         // Filter by delay. The delay is given as a float of hours and must be
         // converted to seconds.
         if let Some(delay) = filters.delay {
-            let max_delay = (delay * 3600.0) as i32;
+            let max_delay = (delay * 3600.0) as i64;
             if let Some(mirror_delay) = mirror.delay {
                 if mirror_delay > max_delay {
                     return false;
